Moved the rockets-fetching code to the state initializer action creator
